agents/hh/requirements.txt
httpx==0.28.1
beautifulsoup4==4.13.4
sqlalchemy==2.0.41
python-dotenv==1.1.0
apscheduler==3.11.0
fastapi==0.115.12
uvicorn==0.34.2
anthropic==0.51.0
